requirements.txt
setuptools
django
djangorestframework
django-cors-headers
drf-yasg
coreapi